+++ lib/presentation/features/intro/load_wallet.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
-import 'package:donationwallet/rust/api/simple.dart';
+import 'package:donationwallet/generated/rust/api/simple.dart';
 import 'package:donationwallet/home.dart';
-import 'package:donationwallet/main.dart';
+import 'package:donationwallet/presentation/states/wallet_state.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
