+++ lib/presentation/features/sending/spend.dart
@@ -1,8 +1,8 @@
-import 'package:donationwallet/rust/api/simple.dart';
-import 'package:donationwallet/global_functions.dart';
-import 'package:donationwallet/main.dart';
-import 'package:donationwallet/outputs.dart';
-import 'package:donationwallet/destination.dart';
+import 'package:donationwallet/generated/rust/api/simple.dart';
+import 'package:donationwallet/presentation/states/wallet_state.dart';
+import 'package:donationwallet/utils/global_functions.dart';
+import 'package:donationwallet/presentation/features/spend/outputs.dart';
+import 'package:donationwallet/presentation/features/spend/destination.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
