+++ lib/presentation/features/spend/outputs.dart
@@ -1,7 +1,8 @@
 import 'package:donationwallet/main.dart';
+import 'package:donationwallet/presentation/states/wallet_state.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:donationwallet/rust/api/simple.dart';
+import 'package:donationwallet/generated/rust/api/simple.dart';
 
 class OutputsScreen extends StatelessWidget {
   const OutputsScreen({super.key});
