+++ lib/presentation/features/wallet/wallet.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 
 import 'package:bitcoin_ui/bitcoin_ui.dart';
-import 'package:donationwallet/global_functions.dart';
-import 'package:donationwallet/main.dart';
-import 'package:donationwallet/spend.dart';
+import 'package:donationwallet/presentation/states/wallet_state.dart';
+import 'package:donationwallet/utils/global_functions.dart';
+import 'package:donationwallet/presentation/features/spend/spend.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:barcode_widget/barcode_widget.dart';
